src/openmas/agent/spade_bdi_agent.py
"""Example integration with spade-bdi library for OpenMAS.

This module demonstrates how to integrate an external BDI library (spade-bdi)
with OpenMAS using the BdiAgent base class.

Note: This is an example implementation and requires the spade-bdi library to be installed.
"""

from typing import Any, Dict, Optional, Type, TypeVar

from openmas.agent.bdi import BdiAgent
from openmas.config import AgentConfig
from openmas.logging import get_logger


# Define a base class for SpadeBDI to use whether the library is available or not
class SpadeBDIAgentBase:
    """Base class to represent the spade_bdi.bdi.BDIAgent class."""

    def __init__(self, *args: Any, **kwargs: Any) -> None:
        pass

    def set_belief(self, *args: Any, **kwargs: Any) -> None:
        pass

    def get_belief(self, *args: Any, **kwargs: Any) -> Optional[Any]:
        return None

    def remove_belief(self, *args: Any, **kwargs: Any) -> None:
        pass

    def get_beliefs(self, *args: Any, **kwargs: Any) -> Dict[str, Any]:
        return {}


# Note: This is for demonstration purposes only
# In a real implementation, you would need to install and import the actual spade-bdi library
try:
    # Type ignore tells mypy to ignore this import
    from spade_bdi.bdi import BDIAgent as ImportedSpadeBDIAgent  # type: ignore

    # Use the imported class as the implementation
    _SpadeBDIImplementation = ImportedSpadeBDIAgent
except ImportError:
    # Use the mock class if import fails
    _SpadeBDIImplementation = SpadeBDIAgentBase

# Create a type alias for use in type annotations
SpadeBDIImplementationType = TypeVar("SpadeBDIImplementationType", bound=SpadeBDIAgentBase)

logger = get_logger(__name__)


class SpadeBdiAgent(BdiAgent):
    """BDI agent implementation using the SPADE-BDI library.

    This class demonstrates how to integrate the SPADE-BDI library with OpenMAS.
    It wraps the BDI functionality provided by SPADE-BDI and maps it to the
    OpenMAS BdiAgent interface.
    """

    def __init__(
        self,
        name: Optional[str] = None,
        config: Optional[Dict[str, Any]] = None,
        config_model: Type[AgentConfig] = AgentConfig,
        env_prefix: str = "",
        asl_file_path: str = "",
    ):
        """Initialize the SPADE-BDI agent.

        Args:
            name: The name of the agent (overrides config)
            config: The agent configuration
            config_model: The configuration model class to use
            env_prefix: Optional prefix for environment variables
            asl_file_path: Path to the AgentSpeak (ASL) file with initial beliefs and plans
        """
        super().__init__(name=name, config=config, config_model=config_model, env_prefix=env_prefix)

        # Store the ASL file path
        self.asl_file_path = asl_file_path

        # SPADE-BDI integration (would be initialized in setup)
        self._spade_bdi_agent: Optional[SpadeBDIAgentBase] = None

        self.logger.info(
            "Initialized SPADE-BDI agent",
            agent_name=self.name,
            asl_file=asl_file_path,
        )

    async def setup(self) -> None:
        """Set up the SPADE-BDI agent.

        This method initializes the SPADE-BDI agent with the ASL file.
        """
        await super().setup()

        # In a real implementation, you would initialize the SPADE-BDI agent
        # self._spade_bdi_agent = SpadeBDIAgentBase(self.name, "password", self.asl_file_path)
        # self._spade_bdi_agent.start()

        self.logger.info("SPADE-BDI agent set up", agent_name=self.name)

    async def update_beliefs(self) -> None:
        """Update the agent's beliefs based on perception.

        This method would integrate with the SPADE-BDI perception system.
        """
        # Example of how you might integrate with SPADE-BDI
        if self._spade_bdi_agent:
            # Synchronize beliefs from OpenMAS to SPADE-BDI
            for belief_name, belief_value in self._beliefs.items():
                # self._spade_bdi_agent.set_belief(belief_name, belief_value)
                pass

    async def deliberate(self) -> None:
        """Run the deliberation cycle.

        In SPADE-BDI, deliberation is handled by the underlying BDI engine.
        This method would synchronize the state after deliberation.
        """
        # SPADE-BDI handles deliberation internally
        # This method could be used to synchronize desires from SPADE-BDI to OpenMAS
        pass

    async def plan(self) -> None:
        """Generate plans for achieving selected desires.

        In SPADE-BDI, planning is handled by the underlying BDI engine.
        This method would synchronize the plans/intentions after planning.
        """
        # SPADE-BDI handles planning internally
        # This method could be used to synchronize intentions from SPADE-BDI to OpenMAS
        pass

    async def execute_intentions(self) -> None:
        """Execute the current intentions.

        In SPADE-BDI, intention execution is handled by the underlying BDI engine.
        This method would trigger or monitor intention execution.
        """
        # SPADE-BDI handles intention execution internally
        pass

    # Override belief methods to integrate with SPADE-BDI
    def add_belief(self, belief_name: str, belief_value: Any) -> None:
        """Add or update a belief, synchronizing with SPADE-BDI.

        Args:
            belief_name: The name of the belief
            belief_value: The value of the belief
        """
        super().add_belief(belief_name, belief_value)

        # Synchronize with SPADE-BDI
        if self._spade_bdi_agent:
            # self._spade_bdi_agent.set_belief(belief_name, belief_value)
            pass

    def remove_belief(self, belief_name: str) -> None:
        """Remove a belief, synchronizing with SPADE-BDI.

        Args:
            belief_name: The name of the belief
        """
        super().remove_belief(belief_name)

        # Synchronize with SPADE-BDI
        if self._spade_bdi_agent:
            # self._spade_bdi_agent.remove_belief(belief_name)
            pass

    def get_belief(self, belief_name: str, default: Any = None) -> Any:
        """Get the value of a belief, synchronizing with SPADE-BDI.

        Args:
            belief_name: The name of the belief
            default: Default value if belief doesn't exist

        Returns:
            The value of the belief or the default value
        """
        # Try to get from SPADE-BDI first
        if self._spade_bdi_agent:
            # spade_value = self._spade_bdi_agent.get_belief(belief_name)
            # if spade_value is not None:
            #     return spade_value
            pass

        # Fall back to OpenMAS beliefs
        return super().get_belief(belief_name, default)

    async def shutdown(self) -> None:
        """Shut down the SPADE-BDI agent.

        This method cleans up the SPADE-BDI agent resources.
        """
        # Clean up SPADE-BDI
        if self._spade_bdi_agent:
            # self._spade_bdi_agent.stop()
            self._spade_bdi_agent = None

        await super().shutdown()
