pyproject.toml
[tool.poetry]
name = "openmas"
version = "0.0.1"
description = "OpenMAS: A lightweight SDK for building Multi-Agent Systems"
authors = ["Wilson <your-email@example.com>"]
readme = "README.md"
packages = [{include = "openmas", from = "src"}]

[tool.poetry.scripts]
openmas = "openmas.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.5.0"
httpx = ">=0.25.0"
structlog = "^23.2.0"
fastapi = "^0.115.12"
pyyaml = "^6.0.1"
typer = "^0.15.2"
python-dotenv = "^1.1.0"
# Optional dependencies
mcp = { version = "1.6.0", optional = true }
grpcio = { version = "^1.71.0", optional = true }
grpcio-tools = { version = "^1.71.0", optional = true }
protobuf = { version = ">=5.26.1,<6.0.0", optional = true }
paho-mqtt = { version = "^1.6.1", optional = true }

[tool.poetry.extras]
mcp = ["mcp"]
grpc = ["grpcio", "grpcio-tools", "protobuf"]
mqtt = ["paho-mqtt"]
all = ["mcp", "grpcio", "grpcio-tools", "protobuf", "paho-mqtt"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
types-requests = "^2.31.0.2"
pre-commit = "^4.2.0"
types-pyyaml = "^6.0.12.20250402"
types-setuptools = "^79.0.0.20250422"
types-paho-mqtt = "^1.6.0.6"
uvicorn = "^0.34.2"
pytest-timeout = "^2.3.1"
tox = "^4.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest]
testpaths = ["tests"]
norecursedirs = ["tests/backup"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.pre_commit]
hooks = [
    "trailing-whitespace",
    "end-of-file-fixer",
    "check-yaml",
    "check-toml",
    "black",
    "isort",
    "flake8",
    "mypy",
    "pytest"
]
